feat(core): Syncing of idw metadata and keria storage (#278)
* build: add connection to keria in agent

* fix: change url oobi correctly

* refactor: clean code and fix ui redux

* build: add unit test and refactor

* fix: remove saved keri record

* fix: refactor code and test

* refactor: get contacts for connections

* refactor: type contacts

* build: show keri connection

* build: add redux update for keri connection

* fix: connection unit test

* build: add keri metadata to connection]

* build: add generate random 21 char key

* test: add further signify unit test

* fix: refactor some comments

* fix: refactor creating keri metadata flow

* fix: refactor coding and unit test

* build: change branch dir to avoid conflict

* build: add all flow insuance credential

* build: update docker-compose

* build: add event publish for acdc

* build: clean the receiving the credential in agent

* fix: change method get aid exchange

* build: add credential pending in UI

* test: adding unit test for ACDC related functions

* fix: unit test

* build: update package lock

* chore: adding create-aid api

* fix: clean the flow with new filter

* chore: rename function

* build: change keria url to remote

* fix: init keri function name

* fix: update package lock file

* build: add connection type to credential details

* test: add unit test for KERI noti not found

* test: add unit test for getCredentialBySaid

* chore: store KERI OOBI in local storage

* chore: upgrade keria & signigy-ts version

* chore: config webpack & jest for the new signify-ts

* chore: get holder using signify

* chore: add create credential and receive connection operations

* feature: get all identifiers, contacts and ACDCs from KERIA

* feat: sync identifiers, contacts and ACDCs from signify

* chore: soft delete credential & identifier

* chore: remove getAllKeriContacts

---------

+++ src/core/agent/services/credentialService.ts
@@ -142,9 +142,12 @@ class CredentialService extends AgentService {
       await this.agent.modules.generalStorage.getAllCredentialMetadata(
         isGetArchive
       );
-    return listMetadatas.map((element: CredentialMetadataRecord) =>
-      this.getCredentialShortDetails(element)
-    );
+    //only get credentials that are not deleted
+    return listMetadatas
+      .filter((item) => !item.isDeleted)
+      .map((element: CredentialMetadataRecord) =>
+        this.getCredentialShortDetails(element)
+      );
   }
 
   getCredentialShortDetails(
@@ -351,7 +354,15 @@ class CredentialService extends AgentService {
   async deleteCredential(id: string): Promise<void> {
     const metadata = await this.getMetadataById(id);
     this.validArchivedCredential(metadata);
-    await this.agent.modules.generalStorage.deleteCredentialMetadata(id);
+    //With KERI, we only soft delete because we need to sync with KERIA. This will prevent re-sync deleted records.
+    if (metadata.connectionType === ConnectionType.KERI) {
+      await this.agent.modules.generalStorage.updateCredentialMetadata(id, {
+        ...metadata,
+        isDeleted: true,
+      });
+    } else {
+      await this.agent.modules.generalStorage.deleteCredentialMetadata(id);
+    }
   }
 
   async restoreCredential(id: string): Promise<void> {
@@ -455,13 +466,19 @@ class CredentialService extends AgentService {
   }
 
   private async createAcdcMetadataRecord(event: any): Promise<void> {
-    const credentialId = event.e.acdc.d;
+    await this.saveAcdcMetadataRecord(event.e.acdc.d, event.e.acdc.a.dt);
+  }
+
+  private async saveAcdcMetadataRecord(
+    credentialId: string,
+    dateTime: string
+  ): Promise<void> {
     const credentialDetails: CredentialShortDetails = {
       id: `metadata:${credentialId}`,
       isArchived: false,
       colors: new ColorGenerator().generateNextColor() as [string, string],
       credentialType: "",
-      issuanceDate: event.e.acdc.a.dt,
+      issuanceDate: dateTime,
       status: CredentialMetadataRecordStatus.PENDING,
       connectionType: ConnectionType.KERI,
     };
@@ -575,6 +592,28 @@ class CredentialService extends AgentService {
       retryTimes++;
     }
     return cred;
+  }
+
+  async syncACDCs() {
+    const signifyCredentials =
+      await this.agent.modules.signify.getCredentials();
+    const storageCredentials =
+      await this.agent.modules.generalStorage.getAllCredentialMetadata();
+    const unSyncedData = signifyCredentials.filter(
+      (credential: any) =>
+        !storageCredentials.find(
+          (item) => credential.sad.d === item.credentialRecordId
+        )
+    );
+    if (unSyncedData.length) {
+      //sync the storage with the signify data
+      for (const credential of unSyncedData) {
+        await this.saveAcdcMetadataRecord(
+          credential.sad.d,
+          credential.sad.a.dt
+        );
+      }
+    }
   }
 }
